test: merge FiatAmount formatted suite into FiatAmountDisplayTests
